Well, it's working now... except
when I press any undefined key. Then, depending on the code, it 1:
loops, even if I press q or something else, 2: closes the app (current
code). I would like it to loop the menu, but run the proper actions when
chosen. In other words, the best of both erros ATM.