Change:
Fixed creation location of PatchIt! settings when the shell extension is
invoked

HEEEEEELLLLPPPP!!!!!!

ARGH! Why will the arguments not work?! Either it short-circuits
preload(), or Args() is not run, creating an error where if the shell is
invoked without the settings being created, thus it throws an error
where the test variable could not be used (as it is right now).